remove need for GCC from makefile
On some platforms gcc is not available, to fix this issue let's use the
cc program which is the system c compiler.

The make variable is now assigned with the '?=' operator, this causes
make to only set the variable if it is not already defined e.g.
`CC=gcc make`